% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-track.R
\name{create_logical_track_filter}
\alias{create_logical_track_filter}
\title{Create a filter for logical track}
\usage{
create_logical_track_filter(
  ltrack,
  filter = NULL,
  filter_name = NULL,
  env = parent.frame()
)
}
\arguments{
\item{ltrack}{name of logical track}

\item{filter}{existing filter (the new filter would be added)}

\item{filter_name}{name for the new filter (optional)}

\item{env}{environment for the new filter (optional)}
}
\value{
a string with the logical track filter (i.e. filter for the values of the original track), added (with '&' operator) to the original filter (if exists).
}
\description{
This function is mostly for internal use. Please create a filter using \code{\link{emr_filter.create}}.
}
\examples{
\dontrun{
create_logical_track_filter("logical_track")
}
}
