% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.emr_expr_vars}
\alias{.emr_expr_vars}
\title{Get a vector of variables inside an expression}
\usage{
.emr_expr_vars(expr)
}
\arguments{
\item{expr}{string with a naryn expression}
}
\value{
vector of the variables inside an expression
}
\description{
Get a vector of variables inside an expression
}
\examples{
.emr_expr_vars("a + b")

}
\keyword{internal}
