% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{.emr_filter}
\alias{.emr_filter}
\title{Parse an R filter string}
\usage{
.emr_filter(filter)
}
\value{
the parsed R filter
}
\description{
This is used internally in the CPP code
}
\examples{
.emr_filter("f > 1")

}
\keyword{internal}
