% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{emr_db.init}
\alias{emr_db.init}
\title{Deprecated - please user emr_db.connect}
\usage{
emr_db.init(
  global.dir = NULL,
  user.dir = NULL,
  global.load.on.demand = T,
  user.load.on.demand = T,
  do.reload = F
)
}
\description{
Deprecated - please user emr_db.connect
}
