% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{emr_db.subset.info}
\alias{emr_db.subset.info}
\title{Returns information about the current subset}
\usage{
emr_db.subset.info()
}
\value{
Information about the current subset or 'NULL'.
}
\description{
Returns information about the current subset.
}
\details{
'emr_db.subset.info' returns the parameters that were used to define the
current subset or 'NULL' if no subset has been defined.
}
\seealso{
\code{\link{emr_db.subset}}, \code{\link{emr_db.subset.ids}}
}
\keyword{~data}
\keyword{~database}
\keyword{~db}
\keyword{~subset}
