% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{emr_db.unload}
\alias{emr_db.unload}
\title{Unload all tracks from naryn database}
\usage{
emr_db.unload()
}
\value{
None.
}
\description{
Unload all tracks from naryn database
}
\examples{
\donttest{
emr_db.unload()
}

}
