% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{emr_download_example_data}
\alias{emr_download_example_data}
\title{Download example database}
\usage{
emr_download_example_data(dir = getwd(), temp_dir = tempdir())
}
\arguments{
\item{dir}{Directory to save the database to. Default: current working directory.}

\item{temp_dir}{Directory to save the temporary downloaded file to.
Change if your system has a small `/tmp`` directory}
}
\value{
None. The database is saved under the name `sample_db` in the specified directory.
}
\description{
Download an example database which was simulated to include an example of a typical EMR database.
}
\examples{
\dontrun{
emr_download_example_data("sample_db")
}

}
