% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.get}
\alias{emr_entries.get}
\title{Get an entry}
\usage{
emr_entries.get(key, db_dir = NULL)
}
\arguments{
\item{key}{The key of the entry to get}

\item{db_dir}{One or more database directories to reload entries from. If NULL - the first database is used.}
}
\value{
The entry value. If the key does not exist, NULL is returned. For multiple databases, a named list of database entries is returned.
}
\description{
Get an entry
}
\examples{
emr_db.init_examples()
emr_entries.get("entry1")

}
