% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.get_all}
\alias{emr_entries.get_all}
\title{Get all entries}
\usage{
emr_entries.get_all(db_dir = NULL)
}
\arguments{
\item{db_dir}{The directory of the database to reload entries from}
}
\value{
A list of entries
}
\description{
Get all entries
}
\examples{
emr_db.init_examples()
emr_entries.get_all()

}
