% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.ls}
\alias{emr_entries.ls}
\title{List entries}
\usage{
emr_entries.ls(db_dir = NULL)
}
\arguments{
\item{db_dir}{The directory of the database to reload entries from}
}
\value{
A list of entries
}
\description{
List entries
}
\examples{
emr_db.init_examples()
emr_entries.ls()

}
