% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.reload}
\alias{emr_entries.reload}
\title{Reload entries from disk}
\usage{
emr_entries.reload(db_dir = NULL)
}
\arguments{
\item{db_dir}{The directory of the database to reload entries from}
}
\value{
None. If the entries were reloaded - the file timestamp is returned invisibly.
}
\description{
Reload entries from disk
}
\examples{
emr_db.init_examples()
emr_entries.reload()

}
