% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.set}
\alias{emr_entries.set}
\title{Set an entry}
\usage{
emr_entries.set(key, value, db_dir = NULL)
}
\arguments{
\item{key}{The key of the entry to set}

\item{value}{The value of the entry to set. This can be anything that can be serialized to YAML}

\item{db_dir}{One or more database directories to reload entries from. If NULL - the first database is used.}
}
\value{
None
}
\description{
Set an entry
}
\examples{
emr_db.init_examples()
emr_entries.set("entry1", "new value")
emr_entries.get("entry1")

}
