% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{emr_filter.clear}
\alias{emr_filter.clear}
\title{Clear all filters from the current environment}
\usage{
emr_filter.clear()
}
\value{
None.
}
\description{
Clear all filters from the current environment
}
\examples{

emr_db.init_examples()
emr_filter.create("f1", "dense_track", time.shift = c(2, 4))
emr_filter.ls()
emr_filter.clear()
emr_filter.ls()
}
