% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{emr_filter.create_from_name}
\alias{emr_filter.create_from_name}
\title{Create a filter from an automatically generated name}
\usage{
emr_filter.create_from_name(filter)
}
\arguments{
\item{filter}{name of a filter automatically generated by \code{emr_filter.name}}
}
\description{
Create a filter from an automatically generated name
}
\examples{
emr_db.init_examples()
name <- emr_filter.name("dense_track", time.shift = c(2, 4))
emr_filter.create_from_name(name)
}
\seealso{
\code{\link{emr_filter.create}}, \code{\link{emr_filter.create_from_name}}
}
\keyword{~filter}
