% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{emr_filter.exists}
\alias{emr_filter.exists}
\title{Checks whether the named filter exists}
\usage{
emr_filter.exists(filter)
}
\arguments{
\item{filter}{filter name}
}
\value{
'TRUE', if the named filter exists, otherwise 'FALSE'.
}
\description{
Checks whether the named filter exists.
}
\details{
This function checks whether the named filter exists.
}
\examples{

emr_db.init_examples()
emr_filter.create("f1", "dense_track", time.shift = c(2, 4))
emr_filter.exists("f1")
}
\seealso{
\code{\link{emr_filter.create}}, \code{\link{emr_filter.ls}}
}
\keyword{~exists}
\keyword{~filter}
