% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{emr_filter.rm}
\alias{emr_filter.rm}
\title{Deletes a named filter}
\usage{
emr_filter.rm(filter)
}
\arguments{
\item{filter}{filter name}
}
\value{
None.
}
\description{
Deletes a named filter.
}
\details{
This function deletes a named filter from current R environment.
}
\examples{

emr_db.init_examples()
emr_filter.create("f1", "dense_track", time.shift = c(2, 4))
emr_filter.create("f2", "dense_track", keepref = T)
emr_filter.ls()
emr_filter.rm("f1")
emr_filter.ls()
}
\seealso{
\code{\link{emr_filter.create}}, \code{\link{emr_filter.ls}}
}
\keyword{~filter}
