% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{emr_filters.info}
\alias{emr_filters.info}
\title{Returns the filter definition of named filters given a filter expression}
\usage{
emr_filters.info(filter)
}
\arguments{
\item{filter}{a filter expression}
}
\value{
a list of named filters
}
\description{
Returns the filter definition of named filters given a filter expression
}
\examples{
emr_db.init_examples()
emr_filter.create("f1", "dense_track", time.shift = c(2, 4))
emr_filter.create("f2", "dense_track", time.shift = c(2, 4))
emr_filter.create("f3", "dense_track", time.shift = c(2, 4))
emr_filters.info("f1 | (f2 & f3)")

}
\seealso{
\code{\link{emr_filter.info}}
}
\keyword{~filter}
