% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.addto}
\alias{emr_track.addto}
\title{Adds new records to a track}
\usage{
emr_track.addto(track, src)
}
\arguments{
\item{track}{track name}

\item{src}{file name or data-frame containing the track records}
}
\value{
None.
}
\description{
Adds new records to a track from a TAB-delimited file or a data frame.
}
\details{
This function adds new records to a track. The records are contained either
in a file or a data frame.

If 'src' is a file name, the latter must be constituted of four columns
separated by spaces or 'TAB' characters: ID, time, reference and value. The
file might contain lines of comments which should start with a '#'
character. Note that the file should not contain a header line.

Alternatively 'src' can be a data frame consisting of the columns named
"id", "time", "ref" and "value". Note: "ref" column in the data frame is
optional.
}
\seealso{
\code{\link{emr_track.import}}, \code{\link{emr_track.create}},
\code{\link{emr_db.init}}, \code{\link{emr_track.ls}}
}
\keyword{~import}
