% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.ids}
\alias{emr_track.ids}
\title{Returns track ids}
\usage{
emr_track.ids(track)
}
\arguments{
\item{track}{track name}
}
\value{
An Ids Table
}
\description{
Returns the ids contained by the track.
}
\details{
Returns the ids contained by the track.

Note: this function ignores the current subset, i.e. ids of the whole track
are returned.
}
\examples{

emr_db.init_examples()
emr_track.ids("categorical_track")
}
\seealso{
\code{\link{emr_track.unique}}, \code{\link{emr_track.info}}
}
\keyword{~ids}
\keyword{~track}
