% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.info}
\alias{emr_track.info}
\title{Returns information about the track}
\usage{
emr_track.info(track)
}
\arguments{
\item{track}{track name}
}
\value{
A list that contains track properties
}
\description{
Returns information about the track.
}
\details{
This function returns information about the track: type, data type, number
of vales, number of unique values, minimal / maximal value, minimal /
maximal id, minimal / maximal time.

Note: this function ignores the current subset, i.e. it is applied to the
whole track.
}
\examples{

emr_db.init_examples()
emr_track.info("sparse_track")
}
\seealso{
\code{\link{emr_track.ls}}
}
\keyword{~info}
\keyword{~property}
\keyword{~track}
