% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-track.R
\name{emr_track.logical.exists}
\alias{emr_track.logical.exists}
\title{Is a track logical}
\usage{
emr_track.logical.exists(track)
}
\arguments{
\item{track}{of the track}
}
\value{
TRUE if \code{track} is a logical track and FALSE otherwise
}
\description{
Is a track logical
}
\examples{
emr_track.logical.exists("logical_track")

}
