% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-track.R
\name{emr_track.logical.info}
\alias{emr_track.logical.info}
\title{Returns information about a logical track}
\usage{
emr_track.logical.info(track)
}
\arguments{
\item{track}{track name}
}
\value{
A list that contains \code{source} - the source of the logical track, and \code{values}: the values of the logical track.
}
\description{
Returns information about a logical track
}
\details{
This function returns the source and values of a logical track
}
\examples{
emr_db.init_examples()
emr_track.logical.info("logical_track")

}
\seealso{
\code{\link{emr_track.ls}}
}
\keyword{~info}
\keyword{~property}
\keyword{~track}
