% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-track.R
\name{emr_track.logical.rm}
\alias{emr_track.logical.rm}
\title{Deletes a logical track}
\usage{
emr_track.logical.rm(track, force = FALSE, rm_vars = TRUE)
}
\arguments{
\item{track}{the name of one or more tracks to delete}

\item{force}{if 'TRUE', suppresses user confirmation of a named track removal}

\item{rm_vars}{remove track variables}
}
\value{
None.
}
\description{
Deletes a logical track
}
\keyword{~create_logical}
\keyword{~track}
