% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.rm}
\alias{emr_track.rm}
\title{Deletes a track}
\usage{
emr_track.rm(track, force = FALSE)
}
\arguments{
\item{track}{one or more track names to delete}

\item{force}{if 'TRUE', suppresses user confirmation of a named track removal}
}
\value{
None.
}
\description{
Deletes a track.
}
\details{
This function deletes a user track from the database. By default
'emr_track.rm' requires the user to interactively confirm the deletion. Set
'force' to 'TRUE' to suppress the user prompt.
}
\seealso{
\code{\link{emr_track.create}}, \code{\link{emr_track.mv}},
\code{\link{emr_track.ls}}, \code{\link{emr_track.readonly}}
}
\keyword{~track}
