% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.unique}
\alias{emr_track.unique}
\title{Returns track values}
\usage{
emr_track.unique(track)
}
\arguments{
\item{track}{track name}
}
\value{
A vector of values
}
\description{
Returns unique and sorted track values
}
\details{
Returns unique and sorted track values. NaN values (if exist in the track)
are not returned.

Note: this function ignores the current subset, i.e. the unique values of
the whole track are returned.
}
\examples{

emr_db.init_examples()
emr_track.unique("categorical_track")

}
\seealso{
\code{\link{emr_track.ids}}, \code{\link{emr_track.info}}
}
\keyword{~track}
\keyword{~unique}
