% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-vars.R
\name{emr_track.var.set}
\alias{emr_track.var.set}
\title{Assigns value to a track variable}
\usage{
emr_track.var.set(track, var, value)
}
\arguments{
\item{track}{track name}

\item{var}{track variable name}

\item{value}{value}
}
\value{
None.
}
\description{
Assigns value to a track variable.
}
\details{
This function creates a track variable and assigns 'value' to it. If the
track variable already exists its value is overwritten.
}
\examples{

emr_db.init_examples()
emr_track.var.set("sparse_track", "test_var", 1:10)
emr_track.var.get("sparse_track", "test_var")
emr_track.var.rm("sparse_track", "test_var")
}
\seealso{
\code{\link{emr_track.var.get}}, \code{\link{emr_track.var.ls}},
\code{\link{emr_track.var.rm}}
}
\keyword{~variable}
