% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrack.R
\name{emr_vtrack.clear}
\alias{emr_vtrack.clear}
\title{Clear all virtual tracks from the current environment}
\usage{
emr_vtrack.clear()
}
\value{
None.
}
\description{
Clear all virtual tracks from the current environment
}
\examples{

emr_db.init_examples()
emr_vtrack.create("vtrack1", "dense_track")
emr_vtrack.ls()
emr_vtrack.clear()
emr_vtrack.ls()
}
