% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrack.R
\name{emr_vtrack.rm}
\alias{emr_vtrack.rm}
\title{Deletes a virtual track}
\usage{
emr_vtrack.rm(vtrack)
}
\arguments{
\item{vtrack}{virtual track name}
}
\value{
None.
}
\description{
Deletes a virtual track.
}
\details{
This function deletes a virtual track from current R environment.
}
\examples{

emr_db.init_examples()
emr_vtrack.create("vtrack1", "dense_track")
emr_vtrack.create("vtrack2", "dense_track")
emr_vtrack.ls()
emr_vtrack.rm("vtrack1")
emr_vtrack.ls()
}
\seealso{
\code{\link{emr_vtrack.create}}, \code{\link{emr_vtrack.ls}}
}
\keyword{~virtual}
