\name{naryn-package}
\alias{naryn-package}
\alias{naryn}
\docType{package}
\title{
Toolkit for medical records data analysis
}
\description{
'naryn' package is intended to help users to efficiently analyze data in time-patient space.

For a complete list of help resources, use \code{library(help = "naryn")}.

More information about the options can be found in 'User manual' of the package.

}
\keyword{ package }
