% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_to_string}
\alias{var_to_string}
\title{Decode a variable created by string_to_var}
\usage{
var_to_string(str)
}
\arguments{
\item{str}{string which was generated by \code{string_to_var}}
}
\value{
the original string
}
\description{
Convert a variable created by \code{string_to_var} back to the original string.
}
\examples{
var_to_string(string_to_var("a & b"))
var_to_string(string_to_var("saba and savta"))
var_to_string(string_to_var("/home/mydir"))
var_to_string(string_to_var("www.google.com"))
var_to_string(string_to_var("my_variable + 3"))
var_to_string(string_to_var(".hidden variable"))
var_to_string(string_to_var("NULL"))

}
