\name{emr_db.reload}
\alias{emr_db.reload}
\title{
Reloads database
}
\description{
  Reloads database
}
\usage{
emr_db.reload()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{
Reloads Naryn Database from disk. Use this function if you manually add tracks or if for any reason the database becomes corrupted.
}

\seealso{
\code{\link{emr_db.init}}, \code{\link{emr_track.ls}}, \code{\link{emr_vtrack.ls}}
}
\keyword{ ~db }
