% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{emr_db.subset.ids}
\alias{emr_db.subset.ids}
\title{Returns the ids that constitute the current ids subset}
\usage{
emr_db.subset.ids()
}
\value{
Ids table or 'NULL'
}
\description{
Returns the ids that constitute the current ids subset.
}
\details{
'emr_db.subset.ids' returns the ids that constitute the current ids subset.
The ids are returned in "ids table" format.

If no ids subset is defined, 'emr_db.subset.ids' returns 'NULL'.
}
\seealso{
\code{\link{emr_db.subset}}
}
\keyword{~data}
\keyword{~database}
\keyword{~db}
\keyword{~subset}
