\name{emr_db.subset.info}
\alias{emr_db.subset.info}
\title{
Returns information about the current subset
}
\description{
Returns information about the current subset.
}
\usage{
emr_db.subset.info()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{
'emr_db.subset.info' returns the parameters that were used to define the current subset or 'NULL' if no subset has been defined.
}
\value{
Information about the current subset or 'NULL'.
}

\seealso{
\code{\link{emr_db.subset}}, \code{\link{emr_db.subset.ids}}
}
\keyword{ ~db }
\keyword{ ~data }
\keyword{ ~database }
\keyword{ ~subset }
