% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.get}
\alias{emr_entries.get}
\title{Get an entry}
\usage{
emr_entries.get(key, db_dir = NULL)
}
\arguments{
\item{key}{The key of the entry to get}

\item{db_dir}{The directory of the database to reload entries from}
}
\value{
The entry. If the key does not exist, NULL is returned.
}
\description{
Get an entry
}
\examples{
emr_db.init_examples()
emr_entries.get("entry1")

}
