% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.get_all}
\alias{emr_entries.get_all}
\title{Get all entries}
\usage{
emr_entries.get_all(db_dir = NULL)
}
\arguments{
\item{db_dir}{One or more database directories to reload entries from. If NULL - the first database is used.}
}
\value{
A list of entries. For multiple databases, a named list of database entries is returned.
}
\description{
Get all entries
}
\examples{
emr_db.init_examples()
emr_entries.get_all()

}
