% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.ls}
\alias{emr_entries.ls}
\title{List entries}
\usage{
emr_entries.ls(db_dir = NULL)
}
\arguments{
\item{db_dir}{One or more database directories to reload entries from. If NULL - the first database is used.}
}
\value{
A vector of entry names. For multiple databases, a named list of database entries is returned.
}
\description{
List entries
}
\examples{
emr_db.init_examples()
emr_entries.ls()

}
