% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.rm}
\alias{emr_entries.rm}
\title{Remove an entry}
\usage{
emr_entries.rm(key, db_dir = NULL)
}
\arguments{
\item{key}{The key of the entry to remove}

\item{db_dir}{The directory of the database to reload entries from}
}
\value{
None
}
\description{
Remove an entry
}
\examples{
emr_db.init_examples()
emr_entries.rm("record1")
emr_entries.ls()

}
