% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{emr_entries.rm_all}
\alias{emr_entries.rm_all}
\title{Remove all entries}
\usage{
emr_entries.rm_all(db_dir = NULL)
}
\arguments{
\item{db_dir}{One or more database directories to reload entries from. If NULL - the first database is used.}
}
\value{
None
}
\description{
Remove all entries
}
\examples{
emr_db.init_examples()
emr_entries.rm_all()

}
