% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{emr_filter.info}
\alias{emr_filter.info}
\title{Returns the definition of a named filter}
\usage{
emr_filter.info(filter)
}
\arguments{
\item{filter}{filter name}
}
\value{
Internal representation of a named filter.
}
\description{
Returns the definition of a named filter.
}
\details{
This function returns the internal representation of a named filter.
}
\examples{

emr_db.init_examples()
emr_filter.create("f1", "dense_track", time.shift = c(2, 4))
emr_filter.info("f1")
}
\seealso{
\code{\link{emr_filter.create}}
}
\keyword{~filter}
