% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{emr_filter.name}
\alias{emr_filter.name}
\title{Generate a default name for a naryn filter}
\usage{
emr_filter.name(
  src,
  keepref = FALSE,
  time.shift = NULL,
  val = NULL,
  expiration = NULL
)
}
\arguments{
\item{src}{source (track name or id-time table)}

\item{keepref}{'TRUE' or 'FALSE'}

\item{time.shift}{time shift and expansion for iterator time}

\item{val}{selected values}

\item{expiration}{expiration period}
}
\value{
a default name for the filter
}
\description{
Generate a default name for a naryn filter
}
\examples{

emr_db.init_examples()
emr_filter.name("dense_track", time.shift = c(2, 4))
}
\seealso{
\code{\link{emr_filter.create}}
}
