% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{emr_traceback}
\alias{emr_traceback}
\title{Prints call stack of the last uncaught error}
\usage{
emr_traceback(x = NULL, max.lines = getOption("deparse.max.lines"))
}
\arguments{
\item{x}{see 'traceback'}

\item{max.lines}{see 'traceback'}
}
\value{
See 'traceback'.
}
\description{
Prints call stack of the last uncaught error in a friendly way.
}
\details{
Similarly to 'traceback' this function prints the call stack of the last
uncaught error. Yet 'emr_traceback' does it in a more friendly way by
omitting the calls that occurred inside the library.
}
\examples{

try({
    emr_db.init_examples()
    f <- function() {
        emr_screen("blablabla")
    }
    f()
    emr_traceback()
})

}
\seealso{
\code{\link{traceback}}
}
\keyword{~error}
\keyword{~exception}
\keyword{~trace}
