% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.attr.get}
\alias{emr_track.attr.get}
\title{Returns the value of the track attribute}
\usage{
emr_track.attr.get(track = NULL, attr = NULL)
}
\arguments{
\item{track}{track name}

\item{attr}{attribute name}
}
\value{
Track attribute value or 'NULL'.
}
\description{
Returns the value of the track attribute.
}
\details{
This function returns the value of a track attribute or 'NULL' if the
attribute does not exist.
}
\examples{

emr_db.init_examples()
emr_track.attr.set("sparse_track", "test_attr", "value")
emr_track.attr.get("sparse_track", "test_attr")
}
\seealso{
\code{\link{emr_track.attr.export}},
\code{\link{emr_track.attr.set}}
}
\keyword{~attr}
\keyword{~attribute}
