% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-attributes.R
\name{emr_track.attr.rm}
\alias{emr_track.attr.rm}
\title{Deletes a track attribute}
\usage{
emr_track.attr.rm(track, attr)
}
\arguments{
\item{track}{track name}

\item{attr}{attribute name}
}
\value{
None.
}
\description{
Deletes a track attribute.
}
\details{
This function deletes a track attribute.
}
\examples{

emr_db.init_examples()
emr_track.attr.set("sparse_track", "test_attr", "value")
emr_track.attr.export()
emr_track.attr.rm("sparse_track", "test_attr")
emr_track.attr.export()
}
\seealso{
\code{\link{emr_track.attr.set}}, \code{\link{emr_track.attr.get}},
\code{\link{emr_track.attr.export}}
}
\keyword{~attr}
\keyword{~attribute}
