% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-attributes.R
\name{emr_track.attr.set}
\alias{emr_track.attr.set}
\title{Assigns a value to the track attribute}
\usage{
emr_track.attr.set(track, attr, value)
}
\arguments{
\item{track}{one or more track names}

\item{attr}{one or more attribute names}

\item{value}{on or more values (strings). Can be an empty string ('').}
}
\value{
None.
}
\description{
Assigns a value to the track attribute.
}
\details{
This function creates a track attribute and assigns 'value' to it. If the
attribute already exists its value is overwritten.

Note that both attributes and values sould be in ASCII encoding.
}
\examples{

emr_db.init_examples()
emr_track.attr.set("sparse_track", "test_attr", "value")
emr_track.attr.get("sparse_track", "test_attr")
}
\seealso{
\code{\link{emr_track.attr.get}}, \code{\link{emr_track.attr.rm}},
\code{\link{emr_track.attr.export}}
}
\keyword{~attr}
\keyword{~attribute}
