% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-track.R
\name{emr_track.create_logical}
\alias{emr_track.create_logical}
\title{Creates a logical track}
\usage{
emr_track.create_logical(track, src, values = NULL)
}
\arguments{
\item{track}{the name of the newly created logical track}

\item{src}{name of the physical track}

\item{values}{vector of selected values}
}
\value{
None.
}
\description{
Creates a logical track
}
\details{
This function creates a logical track based on an existing categorical track
in the global space.
}
\keyword{~create_logical}
\keyword{~track}
