% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.dbs}
\alias{emr_track.dbs}
\title{Returns a vector of db ids which have a
version of the track}
\usage{
emr_track.dbs(track)
}
\arguments{
\item{track}{track name}
}
\value{
A vector of db ids
}
\description{
Returns a vector of db ids which have a
version of the track
}
\examples{

# both db1 and db2 have a track named track1
emr_db.connect(c("/db1", "/db2"))
emr_track.dbs("track1")
}
\seealso{
\code{\link{emr_track.info}}
}
\keyword{~connect}
\keyword{~db}
\keyword{~db_id}
\keyword{~info}
\keyword{~property}
\keyword{~track}
