% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.exists}
\alias{emr_track.exists}
\title{Checks whether the track exists}
\usage{
emr_track.exists(track, db_id = NULL)
}
\arguments{
\item{track}{track name}

\item{db_id}{string of a db dir passed to \code{emr_db.connect}}
}
\value{
'TRUE' if the tracks exists, otherwise 'FALSE'
}
\description{
Checks whether the track exists.
}
\details{
This function checks whether the track exists.
If \code{db_id} is passed, the function checks
whether the track exists in the specific db.
}
\examples{

emr_db.init_examples()
emr_track.exists("sparse_track")
}
\seealso{
\code{\link{emr_track.ls}}, \code{\link{emr_track.info}}
}
\keyword{~exists}
\keyword{~track}
