% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-track.R
\name{emr_track.logical.rm}
\alias{emr_track.logical.rm}
\title{Deletes a logical track}
\usage{
emr_track.logical.rm(track, force = FALSE)
}
\arguments{
\item{track}{the name of the newly created logical track}

\item{force}{if 'TRUE', supresses user confirmation of a named track removal}
}
\value{
None.
}
\description{
Deletes a logical track
}
\keyword{~create_logical}
\keyword{~track}
