% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{emr_track.mv}
\alias{emr_track.mv}
\title{Moves (renames) a track}
\usage{
emr_track.mv(src, tgt, space = NULL)
}
\arguments{
\item{src}{source track name}

\item{tgt}{target track name}

\item{space}{"global" or "user" or 'NULL'}
}
\value{
None.
}
\description{
Moves (renames) a track
}
\details{
This function moves (renames) 'src' track into 'tgt'. If 'space' equals
'NULL', the track remains in the same space. Otherwise it is moved
to the specified space.

Note that logical tracks cannot be moved to the user space.
}
\seealso{
\code{\link{emr_track.create}}, \code{\link{emr_track.rm}},
\code{\link{emr_track.ls}}
}
\keyword{~track}
