% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-vars.R
\name{emr_track.var.get}
\alias{emr_track.var.get}
\title{Returns value of a track variable}
\usage{
emr_track.var.get(track, var)
}
\arguments{
\item{track}{track name}

\item{var}{track variable name}
}
\value{
Track variable value. If the variable does not exists, NULL is returned.
}
\description{
Returns value of a track variable.
}
\details{
This function returns the value of a track variable. If the variable does
not exist NULL is returned.
}
\examples{

emr_db.init_examples()
emr_track.var.set("sparse_track", "test_var", 1:10)
emr_track.var.get("sparse_track", "test_var")
emr_track.var.rm("sparse_track", "test_var")
}
\seealso{
\code{\link{emr_track.var.set}}, \code{\link{emr_track.var.ls}},
\code{\link{emr_track.var.rm}}
}
\keyword{~variable}
