% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrack.R
\name{emr_vtrack.exists}
\alias{emr_vtrack.exists}
\title{Checks whether the virtual track exists}
\usage{
emr_vtrack.exists(vtrack)
}
\arguments{
\item{vtrack}{virtual track name}
}
\value{
'TRUE' if the virtual track exists, otherwise 'FALSE'.
}
\description{
Checks whether the virtual track exists.
}
\details{
This function checks whether the virtual track exists.
}
\examples{

emr_db.init_examples()
emr_vtrack.create("vtrack1", "dense_track", time.shift = c(5, 10), func = "max")
emr_vtrack.exists("vtrack1")
}
\seealso{
\code{\link{emr_vtrack.create}}, \code{\link{emr_vtrack.ls}}
}
\keyword{~exists}
\keyword{~virtual}
