% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrack.R
\name{emr_vtrack.info}
\alias{emr_vtrack.info}
\title{Returns the definition of a virtual track}
\usage{
emr_vtrack.info(vtrack)
}
\arguments{
\item{vtrack}{virtual track name}
}
\value{
Internal representation of a virtual track.
}
\description{
Returns the definition of a virtual track.
}
\details{
This function returns the internal represenation of a virtual track.
}
\examples{

emr_db.init_examples()
emr_vtrack.create("vtrack1", "dense_track", "max", time.shift = c(5, 10))
emr_vtrack.info("vtrack1")
}
\seealso{
\code{\link{emr_vtrack.create}}
}
\keyword{~virtual}
