% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naryn-package.R
\docType{package}
\name{naryn-package}
\alias{naryn-package}
\alias{naryn}
\title{Toolkit for medical records data analysis}
\description{
'naryn' package is intended to help users to efficiently analyze data in
time-patient space.
}
\details{
For a complete list of help resources, use \code{library(help = "naryn")}.

More information about the options can be found in 'User manual' of the
package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tanaylab.github.io/naryn/}
  \item Report bugs at \url{https://github.com/tanaylab/naryn/issues}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Authors:
\itemize{
  \item Misha Hoichman \email{misha@hoichman.com}
  \item Ben Gilat \email{ben.gilat@weizmann.ac.il}
}

Other contributors:
\itemize{
  \item Netta Mendelson-Cohen \email{Netta.Mendelsoncohen@weizmann.ac.il} [contributor]
  \item Rami Jaschek \email{rami.jaschek@weizmann.ac.il} [contributor]
  \item Weizmann Institute of Science [copyright holder]
}

}
\keyword{package}
